Refactor wallet and seed phrase logic for local or deployed test envs, refactor parallel flag logic
