+++ .env.local.e2e
@@ -1,5 +1,6 @@
 PARALLEL=true
 DEV_MODE=true
+LOCAL_NODE=true
 RPC_URL=http://localhost:9944
 WEBSOCKET=ws://localhost:9944
 ANVIL=http://localhost:8545
