+++ .env.local.rpc
@@ -1,11 +1,7 @@
 PARALLEL=false
+LOCAL_NODE=true
 DEV_MODE=true
 RPC_URL=http://locahost:9944
 WEBSOCKET=ws://127.0.0.1:9944
 BLOCK_HASH=0x7b02642cc4c2012f101256e3a5b3e69f0b61799be0c955fd08042c4bf9d04d9d
 SEED_PHRASE_1=//Alice
-SEED_PHRASE_2='INSERT_SEED_PHRASE'
-SEED_PHRASE_3='INSERT_SEED_PHRASE'
-SEED_PHRASE_4='INSERT_SEED_PHRASE'
-SEED_PHRASE_5='INSERT_SEED_PHRASE'
-SEED_PHRASE_6='INSERT_SEED_PHRASE'
