+++ .env.testnet.rpc
@@ -1,5 +1,6 @@
 PARALLEL=true
 DEV_MODE=false
+LOCAL_NODE=false
 RPC_URL=https://testnet-rpc.zkverify.io
 WEBSOCKET=wss://testnet-rpc.zkverify.io
 BLOCK_HASH=0x7b02642cc4c2012f101256e3a5b3e69f0b61799be0c955fd08042c4bf9d04d9d
