+++ src/e2e-tests/tests/verify/index.test.ts
@@ -1,26 +1,26 @@
 import { runProofWithoutAttestation, runProofWithAttestation, runVKRegistrationTests } from '../../../tests';
-import { fundLocalE2eTestWallets } from '../../../utils/helpers';
+import { setupLocalOrExistingWallets } from '../../../utils/wallets';
 
 jest.setTimeout(300000);
 
+let isParallel: boolean;
+
 beforeAll(async () => {
-    const fundingSuccess = await fundLocalE2eTestWallets();
-    if (!fundingSuccess) {
-        throw new Error('Funding failed. Cannot proceed with tests.');
-    }
+    await setupLocalOrExistingWallets();
+
+    isParallel = process.env.PARALLEL?.toLowerCase() === 'true';
 });
 
 describe('zkVerify proof journey tests', () => {
     test('should verify all proofs without waiting for a published attestation', async () => {
-        await runProofWithoutAttestation(process.env.PARALLEL === 'true');
+        await runProofWithoutAttestation(isParallel);
     });
 
     test('should verify all proofs, called poe pallet and wait for published attestation', async () => {
-        await runProofWithAttestation(process.env.PARALLEL === 'true');
+        await runProofWithAttestation(isParallel);
     });
 
     test('should register VK and verify a proof using the VK hash', async () => {
-        await runVKRegistrationTests(process.env.PARALLEL === 'true');
+        await runVKRegistrationTests(isParallel);
     });
 });
-
